# Using curl_mime_data_cb to capture HTTP send/recv times for TorchServe