--- a/src/clients/c++/perf_analyzer/client_backend/torchserve/http_client.cc
+++ b/src/clients/c++/perf_analyzer/client_backend/torchserve/http_client.cc
@@ -119,7 +119,6 @@ HttpInferRequest::FileSize()
 Error
 HttpInferRequest::CloseFileData()
 {
-  fclose(file_ptr_.get());
   file_ptr_.reset(nullptr);
   return Error::Success;
 }