--- a/src/clients/c++/perf_analyzer/client_backend/torchserve/http_client.h
+++ b/src/clients/c++/perf_analyzer/client_backend/torchserve/http_client.h
@@ -117,7 +117,9 @@ class HttpInferRequest {
   struct Deleter {
     void operator()(FILE* file)
     {
-      // Do nothing
+      if (file != nullptr) {
+        fclose(file);
+      }
     }
   };